service: Move mii enums and structs into its own file
Moves these into types.h, since other files also make use of these types.
